requirements.txt
pillow==10.2.0
matplotlib==3.8.3
numpy==1.26.4